dags/requirements.txt
apache-airflow-providers-google==10.1.0
apache-beam==2.59.0
